requirements.txt
dacite==1.7.0
snowflake-connector-python==3.1.0
snowflake-snowpark-python==1.6.1
